updated profile test to redirects after create profile, added mocks folder
