+++ src/mocks/resolvers.js
@@ -1,0 +1,12 @@
+const mockUser = {
+  userId: '0',
+  username: 'mockuser',
+};
+
+const mockProfile = {
+  ...mockUser,
+  bio: 'mock bio',
+  avatar: 'mock url',
+};
+
+export { mockUser, mockProfile };
