+++ src/views/Profile/Profile.test.jsx
@@ -3,20 +3,10 @@ import userEvent from '@testing-library/user-event';
 import { MemoryRouter, Route, Routes } from 'react-router-dom';
 import { rest } from 'msw';
 import { setupServer } from 'msw/node';
+import { mockUser, mockProfile } from '../../mocks/resolvers';
 import { UserProvider } from '../../context/UserContext';
 import Profile from './Profile';
 import CreateProfile from '../CreateProfile/CreateProfile';
-
-const mockUser = {
-  userId: '0',
-  username: 'mockuser',
-};
-
-const mockProfile = {
-  ...mockUser,
-  bio: 'mock bio',
-  avatar: 'mock url',
-};
 
 const server = setupServer(
   rest.get(`${process.env.API_URL}/api/v1/users/me`, (req, res, ctx) => {
@@ -24,7 +14,7 @@ const server = setupServer(
   }),
 
   rest.get(
-    `${process.env.API_URL}/api/v1/profiles/${mockUser.username}`,
+    `${process.env.API_URL}/api/v1/profiles/:username`,
     (req, res, ctx) => {
       return res(ctx.json({ message: 'this user has no profile' }));
     }
@@ -32,7 +22,14 @@ const server = setupServer(
 
   rest.post(`${process.env.API_URL}/api/v1/profiles`, (req, res, ctx) => {
     return res(ctx.json(mockProfile));
-  })
+  }),
+
+  rest.get(
+    `${process.env.API_URL}/api/v1/projects/user/:user_id`,
+    (req, res, ctx) => {
+      return res(ctx.json({ message: 'this user has no projects' }));
+    }
+  )
 );
 
 describe('Profile', () => {
@@ -64,11 +61,33 @@ describe('Profile', () => {
 
     await user.click(createProfileButton);
 
-    // If navigated to /user/new, should display:
-    await screen.findByRole('textbox', { name: /bio/i });
-    await screen.findByRole('textbox', { name: /avatar/i });
-    await screen.findByRole('button', {
+    // REDIRECT:
+    //    If navigated to /user/new, should display the following:
+    const bioInput = await screen.findByRole('textbox', { name: /bio/i });
+    const avatarInput = await screen.findByRole('textbox', { name: /avatar/i });
+    const saveProfileButton = await screen.findByRole('button', {
       name: /create/i,
     });
+
+    await user.type(bioInput, 'mock bio');
+    await user.type(avatarInput, 'mock image url');
+
+    await user.click(saveProfileButton);
+
+    // REDIRECT:
+    //    On redirect to user profile, response carries new user profile data
+    server.use(
+      rest.get(
+        `${process.env.API_URL}/api/v1/profiles/:username`,
+        (req, res, ctx) => {
+          return res(ctx.json(mockProfile));
+        }
+      )
+    );
+
+    await screen.findByRole('heading', { name: /mockuser/i });
+    await screen.findByAltText(/user avatar/i);
+
+    screen.debug();
   });
 });
